refactor: Remove docstrings and relocate data class
