+++ aws_lambda_powertools/utilities/data_classes/appsync_resolver_event.py
